Reorganize portal stories into separate files for default and features.
